backend/requirements.txt
fastapi
uvicorn
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
python-dotenv
python-multipart
PyPDF2
newspaper4k
pandas
